INKLB-008:REF:database:adapter implementation in header
