+++ backend/utils/db_adapter/include/idb_adapter.h
@@ -6,7 +6,6 @@
 #include <string>
 #include <vector>
 
-
 namespace inklink::db_adapter
 {
 template <typename Implementation, typename Connection>
@@ -17,19 +16,39 @@ public:
     using DbTable = std::vector<DbRow>;
 
 public:
-    void Connect(const std::string& connectionString);
-    [[nodiscard]] std::shared_ptr<Connection> GetConnection();
+    void Connect(const std::string& connectionString)
+    {
+        static_cast<Implementation*>(this)->Connect(connectionString);
+    }
+
+    [[nodiscard]] std::shared_ptr<Connection> GetConnection()
+    {
+        static_cast<Implementation*>(this)->GetConnection();
+    }
+
 
     template <typename... Arguments>
-    void Insert(const std::string& request, const Arguments&... arguments) const;
+    void Insert(const std::string& request, const Arguments&... arguments) const
+    {
+        static_cast<Implementation*>(this)->Insert(request, arguments...);
+    }
 
     template <typename... Arguments>
-    void Update(const std::string& request, const Arguments&... arguments) const;
+    void Update(const std::string& request, const Arguments&... arguments) const
+    {
+        static_cast<Implementation*>(this)->Update(request, arguments...);
+    }
 
     template <typename... Arguments>
-    [[nodiscard]] DbTable Select(const std::string& request, const Arguments&... arguments) const;
+    [[nodiscard]] DbTable Select(const std::string& request, const Arguments&... arguments) const
+    {
+        static_cast<Implementation*>(this)->Select(request, arguments...);
+    }
 
-    void Delete(const std::string& request) const;
+    void Delete(const std::string& request) const
+    {
+        static_cast<Implementation*>(this)->Delete(request);
+    }
 
 protected:
     std::shared_ptr<Connection> m_connection;
